feat: Implement ThankYouCard widget and update navigation for registration success
